Created a basic GUI for combat with non functional buttons.
Created a UI document Which holds a combatUI uxml. When the fight button is pressed the option buttons are replaced by a moves list. The text on the buttons is pulled from the Moncarg instance attached to the UI document. Also created a combatUIManager c# class to handle the button inputs.