Combat UI integration with Combat Handler completed
moncarg health and mana labels added as a private attribute with public setter.

combat handler is integrated with the CombatUI UI document.

Encounter Loop has been refactored to event-based, instead of polling-based to register and execute button presses.

Right now the combat handler and UIDocument variables are in Moncarg so we can assign them in the inspector, remember to move those to Game Manager once we merge with main