Gave Moncarg data appropriate skills.  Made a broken switch button
I gave all the moncargs skills relevant to their type. Then i started work on the switch function. Currently the switch function is immensely buggy as i tried reusing the moncarg selction ui that is used at the beginning of a fight however this creates duplicates of the moncargs in the scene and calls "beginEncounter" multiple times.