Made the switch function work correctly
The switch button now works without creating duplicates and the stats are saved when a moncarg is switched out so that when it is switched in again it will  retain its stats. I did this by commenting out two lines in the moncarg classes function initstats() where the current hp and mana were being set to their maximum.