player movement between rooms
upon entering a door, the player will now move to a new room (which is consistent with the map)

however, the spawn position is fixed and need to be changed based on where the door is. Right now if the player spawns into a collission tile, they will be stuck and be unable to move.