Basic combat terminal still needs work
Created a basic combat terminal and linked it to the various event triggers. Currently it can displayed up to 3 events in order but it appears it may need to show up to 6 so that the player can see what happened last turn as well as whats happening now. There are a few text based bugs ill also iron out before next commit.