FIxed music levels. Icons. and being able to play post victory
Made it so that the combat music is louder when its played to match the level of the background music. change all the used sprites to single instead of multiple so icon now holds the entire image instead of a section of it. made it so that once wictory is achieved when you return to the main menu it automatically starts a new game.