save function should work on android too + bug fixes
add main menu button to victory and lose screen too.
first moncarg auto selected for selection screen so player does not have to click away and click back again

changed save data path to Application.persistentDatapath in stead of Application.dataPath so that it works on android too.

On windows, saves files will be saved to: \Users\<username>\AppData\LocalLow\<CompanyName>\<ProductName>\SaveData\

On android, it will be saved to:
/storage/emulated/0/Android/data/<package_name>/files/SaveData/